Heavily simplify styling architecture
There is no longer a build artifact for the styling, we export scss and leave the consumer to decide what to do with it. This enabled us to simplify the way we expose variables and mixins, and there is no longer an issue with extending base Bootstrap classes through scss `@extend`. As a result `@42.nl/ui-core-styling` has been merged in `@42.nl/ui`.