dx: added Verdaccio local publishing to test packages locally.
Verdaccio is a "A lightweight private npm proxy registry". See:
https://www.verdaccio.org/

To publish to the Verdaccio registry added a script called `publish.sh`
with the environment variable `dev` it will start up Verdaccio in docker
and publish the packages.

In order to be able to publish packages they need a new version. To
always create a completely new version two scripts were written:
`rewrite-versions.js` and `restore-versions.js`. One script to change
all versions and dependencies to a `current date` based version. The
other to restore the versions to what they previously were.