Improvement: Fixed failing tests after greenkeeper update.
Due to the update of `testing-library/react-hooks` the way
`useSettledErrors` was tested had to change. Also had to bump react
itself to `16.9.0`.

Also fixed minor bug in `ModalPickerMultiple` which caused the `selected`
and `value` to be the same, this caused cancelling not to have any effect.
Also fixed storys so the value is actually selected in storybook by
using `useState` and providing the value.