Improvement: dumped down the FlashMessage component.
Before the `FlashMessage` component used the `@42.nl/react-flash-messages`
library to render and close the flash messages. This is to much knowledge
of the way flash messages work. Now the flash message is simply a UI
component which does not know anything about `@42.nl/react-flash-messages`.

The styling has also been changed: it now renders in the full width of
the given container, and it now longer renders on a fixed position.
This is done because the `FlashMessage` no longer knows were it is
rendered anymore.

Users of `@42.nl/ui` should now implement their own flash message engine
or simply use: ``@42.nl/react-flash-messages`.

Also exposed `Color` in the default export.

Closes #78