Feature: ValuePicker and CheckboxMultipleSelect
Added the `ValuePicker` component, which is an abstraction which automatically
selects the best component to use when the user has to select a value
from a pre-defined list.

This is the decision matrix:

```
| multiple | items <= 10             | items > 10
| true     | CheckboxMultipleSelect  | ModalPickerMultiple
| false    | Select                  | ModalPickerSingle
```

`ValuePicker` starts in a booting state in which a spinner is shown.
During the booting state it will call `FetchOptionsCallback` and ask
for a `Page` of size `1` so it can get the `totalElements`.

Refactored common typings to the `src/form/types.ts` so they can be
re-used, and to make the `ValuePicker` work.

Changed the  `FetchOptionsCallback` callback so it accepts a third
parameter called  `size`, so `ValuePicker` can very efficiently get the
`totalElements`. Plus it allows components to defined the size they can
best display. Currently all sizes are `10`.

Added a new component called `CheckboxMultipleSelect` which renders
multiple checkboxes side by side from which the user can choose.

The `Select` component now more intelligently selects a default value
after it has loaded them when options is an `OptionsFetcher`. Before
it would always select the first value which was loaded. Now it only
selects the first value if the current value is not in the loaded
`Page` or when the current value is empty. This had to be done so
switching between `ModalPickerSingle` and `Select` in the `ValuePicker`
would work dynamically.

Also fixed minor UI glitch which caused to `ModalPickerSingle`'s button
not to align when no option was selected. Fixed by only showing the
`span` containing the `value` when a value is present.

Also typed several test utility functions so TypeScript understands them.

Closes: #67 and #66.