Improvement: expand peer dependencies
Added more libraries which are common to 42 projects to the
`peerDependencies` list as well as included them to the `devDependencies`
list. This is due to issues with out-of-sync dependencies being used by
`@42.nl/ui` in conjunction to most of the internal 42 projects.

In addition this commit also resolves most of the `peerDependency`
warnings of the included dependencies.