Improvement: `AsyncContent` now shows retry button on failure.
When the `retry` button is `clicked` a call to `useAsync` return value
`reload` is called. By default this functionality is enabled but can
be disabled by setting `showRetryButton` to `false`.

Added better jsdoc documentation as well.

Also the title for the `loading` and `error` state can now be overridden
by the user.

Closes: #68