feat: Added SearchInput component.
The `SearchInput` component is a small wrapper around the `reactstrap`
`Input` component which debounces the `onChange` by `500` milliseconds
by default. It uses `lodash`'s `debounce` under the hood.

Features:

  1. When the user presses enter the `SearchInput` will skip the debounce
     and set the value straight away.

  2. The icon which is a magnifying class can be toggled on or of.

  3. The debounce timeout can also be set to a custom value.

  4. The debounce settings which lodash accepts can also be set for
     maximum configurability.

The modal pickers now use the `SearchInput` instead of debouncing the
inputs themselves. `ModalPickerMultiple` now uses `Array.isArray` instead
of `lodash.isArray`.

Also had to set the border color in `.form-control:focus` to a gray
to prevent it from becoming blue for the `SearchInput`.

Closes #191