bug: The `NavigationItem` no longer forces `exact links
Sometimes we use nested routes in applications, which do no work
nicely with the default `exact` property. For example when a user navigates
to `users/10` and we have a master-detail view, the `NavigationItem` will
not apply active to the entry, as it has an `exact` property by default.

This commit makes the `exact` configurable by the user to bypass this
behavior.

Fixes #120