improvement: SearchInput's value can now be manipulated from the outside.
The `SearchInput` is debounced and therefore must use an uncontrolled
input. Otherwise the value of the input cannot be debounced.

This causes a problem however, when ever the `value` prop of the
`SearchInput` changed due to external changes. For example: a button
which clears the `SearchInput`. The new value would not appear in the
input element, due to it being uncontrolled.

This commit adds a new children prop to the `SearchInput`. The children
is expected to be a function of type:

```ts
(searchInput: React.ReactNode, api: SearchInputApi) => React.ReactNode;
``

It should take the searchInput rendered by the `SearchInput` and display
it somewhere in the return of the children function. It can the use the
`api` which contains a function called `setValue` to change the actual
value of the `SearchInput`.

The `setValue` will cancel the debounce, update the `<input>`'s actual
value through the ref, and it will call `props.onChange` with the actual
value.

Closes: #219