feat: Added IconPicker form component.
With the `IconPicker` form component the user can select an icon from
the material design icons.

Changed the icons to an array of strings, and create the `IconType` type
from that array of strings. This way we can iterate over the icons and
render them in the IconPicker.

Also removed icons which did not exist in our version of the material icons.
Bumped the material icons to the latest version which is `v49`.

Promoted the `pageOf` test utility to the actual source code. It is used
in the `IconPicker` to create pages from the icons array.

Minor change: fixed jarb comment in `CheckboxMultipleSelect` and formatted
the file.

Closes #208