fix: enabled `strictNullChecks` in tsconfig. Fixed crashing <Icon>.
The `Icon` when no `onClick` prop was provided would crash when clicked.
This was caused because when `onClick` is not a function calling it
results in an error.

This is strange because TypeScript should have caught this error. When
`onClick` can be `undefined` or a `function` it should warn us when we
have not made sure that it is a function.

Turns out that when `strictNullChecks` is `false`, which is the default,
any destructuring makes that type a non-null. This is of course wrong!

So this commit enables the `strictNullChecks` compiler flag. Also adds a
`tsc` script which checks the entire code. This way it is easier to get
all TypeScript errors out of the way first, whether they are in a story
or in a test or in the source code.

Enabling this flag causes a lot of errors:

  1. Added @ts-ignore on tests which called `ShallowWrapper.find` and
     proceed as if there is always something found. If they find nothing
     the test will fail anyway.

  2. Typed the `EpicTable` since it now misses a lot of types. A number
     checks are now written to ensure non null.

  3. ValuePicker components no longer say they can emit `null` on `onChange`
     since they do not do that.

  4. TypeaheadMultiple and TypeaheadSingle emit `undefined` instead of `null`
     when the value is empty. This makes for an easier type because the `value`
     can always be `undefined` already.

  5. ImageUpload removed `defaultProps.text` as empty object construction
     which `TypeScript` does not understand now. Added tests for custom
     text to prove it still works.

  6. Added extra `onClick` exists check on `ProgressStepper`.

  7. Made explicit various variable definitions, so TypeScript compiles.

Tiny refactor: changed ImgUpload to ImageUpload in the test.

Fixes: #222