improvement: `withJarb` will now throw errors when detecting wrong props.
Before this commit when the user accidentally used one of the props
that `withJarb` managed for you. The form element would no longer work.

Unfortunately I cannot get `TypeScript` to understand that there is
a set of all props, from JarbField and from the props of the wrapped
component. And that it should strip away the managed props. I've tried
for 2 days and I cannot get this to work.

So now we simply throw errors when seeing that the user has provided
managed properties. A very detailed error message instructs the user
to trust `withJarb` and to remove those properties.

Closes: #228