improvement: EpicTable is now flat when no horizontal scrollbar is shown
Multiple people are taken aback by the shadows produced on a very small
table. This is because they do not see that the center can be scrolled
in this case. So when a table is small enough not to need a horizontal
scrollbar it should not render the shadows and instead be more flat.

The technical solution is to calculate the total desired width of the
center as given by the user. When the desired width can be met, there is
no need for a horizontal scrollbar. In that case we do not render shadows
around the left and right column, but only around the container.

Also the `GooeyCenter` now renders at 100% width to prevent width
flickering.

Closes #230