bug: `EpicTable` no longer flashes when moving to the flat mode.
When the `EpicTable` switched flat and normal the shadows shown by normal
mode would show very briefly.

The reason for this is because the `EpicTable` calculates in which mode
it needs to render by checking if the desired width requirements have
been met. It can only do this by actually rendering itself. Initially
it will assume the `normal` render, if everything fits it will render
in `flat` mode. This switching between them causes a render flash.

The solution is setting the CSS property `visibility: hidden;` on the
`EpicTable` and on the `GooeyCenter`'s center div. With `visibility: hidden;`
the browser will not show the element but does let it take up the space.
This way we do not shown anything until the calculation has been made.
This of course prevents any flickering.

Fixes: #240