fix: EpicTable's expander / detail row were rendered under the overlay.
When the table is completely empty is is common to show a `ContentState`
telling the user that the table is empty. However clicking on a "create"
button to show a form in the detail row would not work, because the
DetailRow was rendered under the overlay, making it invisible.

The fix was to simply put a higher `z-index` on the detail and expander
row than the overlay has.

Fixes #245.