fix: FixedHeader was broken and did not render correctly.
The reason was a regression in `e9e2693bb` which changed the way the
`EpicLayout` was created. Since that commit an empty header was accidentally
always added on the left.

This causes the arrays length for the center, right and left to diverge.
Since the `FixedHeader` looks at the `left` to determine which row
index it should make fixed, when these arrays vary you get null pointers.

To fix this I've done the following:

  1. In the `layout` the left now has the same guards against empty sections
     and rows as the right and center have. This way we prevent empty
     sections and rows from existing.

  2. The `FixedHeader` when shown would show flashes. Ironically these
     flashes were caused by the guards against flashing. The `GooeyCenter`
     is invisible until the `centerWidth` has been calculated. This
     works for all non `FixedHeader` `GooeyCenter`'s', but not for the
     `FixedHeader`.

     So added a boolean prop called `forceCenterAlwaysVisible` on the
     `GooeyCenter`. The `FixedHeader` sets this to `true` and this way
     there is no flash, for the `FixedHeader`.

  3. Added guards in the `FixedHeader` against causing null pointers.
     This way when something goes wrong at least the application can
     still be used, even though the `FixedHeader` will probably look
     strange.

  4. All stories are written now in a way that promotes triggering
     the FixedHeader. Before the content was often to small to trigger
     the FixedHeader, making these bugs harder to catch.

Also removed the shadows in the `FixedHeader` which looked strange.

Made a tiny refactoring in the code which determines if shadows should
be rendered in the `EpicTable`. Now they are defined in `const`'s for
re-use.

Fixes: #247