improvement: all ValuePickers support custom `isOptionEqual` checks now.
Before the equality of selections were determined based on if the two
values had the same label. This is a good heuristic for the default
implementation, which it still is.

Now however the user can supply a custom checker via the `isOptionEqual`
prop. Here the user implements a feature that gives two values and must
return a boolean. This allows the user to have full control of the
check whether a value is selected or not.

For example you could implement `isOptionEqual` like this to check the
`id` of the items.

```tsx
<ValuePicker
  {/* Rest of the props omitted for brevity */}
  isOptionEqual={(a: User, b: User) => a.id === b.id}
/>
```

You could even use `lodash.isEqual` when you really want a deep check
for equality:

```tsx
<ValuePicker
  {/* Rest of the props omitted for brevity */}
  isOptionEqual={lodash.isEqual}
/>
```

Closes: #209