fix: SearchInput did not update the `onChange` callback.
Because the `onChange` callback is wrapped inside of a `useRef` which
in turn calls `lodash.debounce`. The `useRef` will not update anything
when the `useRef` functions params change.

We must do this manually, added an `useEffect` which updates the
`useRef.current` whenever the params to the `debounce` change.

This way a new `onChange` callback is actually used.

Closes: #265