fix: `EpicTable` did not grow with the `EpicDetailRow`.
This had the effect that when the `EpicDetailRow` is bigger than the
`EpicTable` it would become bigger than the `EpicTable` itself. Possible
overflowing other elements on the page. Which looks strange.

To fix this the `EpicTable`'s height is now based on three things:

  1. The minHeight props of the EpicTable.
  2. The height of the EpicDetailRow.
  3. The desired height of the EpicTable, which is the height of one row.

Which ever height is the largest height becomes the actual height of
the `EpicTable`.

Whenever an EpicDetailRow is active it will adjust its height to the
actual height of the `EpicTable`

Fixes: #269