feat: added `RadioGroup` and `Loading` component.
The `RadioGroup` is the twin brother of the `Select` component only
it renders `<input type='radio' />` instead of a `<select/>` Due to
the overlap between the `Select` and the `RadioGroup` I've created a
hook for their common functionality called: `useOptions`.

The `useOptions` hook loads the options if needed and makes sure
that the selected value is always present in the `options` array.
This add-when-not-exists-behavior fixes #246.

Also created the `Loading` component which used to be part of the
`<Select>` component. It is a tiny loading component which is now
also used by the `RadioGroup`. Also exported it for use by our users.

The `ValuePicker` now uses the `RadioGroup` when in single mode and the
elements are less than 4.

Changed the jest config so it does not try to generate coverage from
the '_site' directory anymore by setting the rootdir to src.

Also moved `test` only sources outside of the `src` folder and into
the root directory. This way all code in `src` can be considered code
which should be packaged.

Closes: #270