fix: form example strange text-area resizing bug.
Because the `Col` of the `FinalForm` component did not have a size
constraints it will by default try do divide the width evenly.

Because the `<pre>` element gets really large when the user enters a
large value into the `Textarea` the division of width is unfairly given
to the right `Col`.

Fixed by adding a `lg={6}` to the `Col`'s making the widths explicit.
This prevents `bootstrap` from dividing the width itself.

Fixes: #80