improvement: `AsyncContent` now supports an empty state.
The `AsyncContent` now has an `isEmpty` callback, through which it
will ask the user if he / she considers the fetched data empty. If
so it will render a `ContentState` in the `empty` mode.

Also added `AsyncPage` and `AsyncList` components.

The `AsyncList` is a variant of the `AsyncContent` which expects
the result of a call to `useAsync` from `react-async` to be an
array.

It has all the behaviors of the `AsyncContent` but provides a
standard `isEmpty` function which checks if the array has the
`length` zero.

The `AsyncPage` is a variant of `AsyncContent` which expects the
result of a call to `useAsync` from `react-async` to be a `Page`
from `@42.nl/spring-connect`.

It has all the behaviors of the `AsyncContent` but provides a
standard `isEmpty` function which checks if the `totalElements`
of the `Page` is zero.

Closes: #283