fix: EpicTable did not resize when EpicDetail / overlay became bigger.
This happens when the `EpicDetail` displays content it loads
asynchronously from the back-end. After the content is loaded the
`EpicDetail` can becomes larger than the loading state quite easily.
This could also happen with an overlay as well.

This results in the `EpicDetail` rendering over the `EpicTable`'s
bounds.

With this fix the `useCalculateActualHeight` now takes into account the
size of the overlay and the EpicDetail.

For the `EpicDetail` an interval is introduced which at every 20
milliseconds checks if the `EpicDetail` has become bigger. This way we
spot changes in size quickly.

Unfortunately it is not possible to make the `EpicDetail` shrink when
it becomes smaller. This is because we set the `style.height` of the
detail manually which makes the `scrollHeight` unusable because we
override it. But this at least does not look as bad as when the
`EpicDetail` is rendered outside of the `EpicTable`.

Fixes #298.