Added optional label to SearchInput
You might want to display a label above the SearchInput to clarify which
SearchInput is used for what purpose in cases where multiple SearchInput
are used inside the same frame/component.
Added possibility to specify a label to display above the input field.
Wraps the field in a FromGroup when label is specified.

Closes #336