Fix: Defined size as property of SearchInput
Size is used when showIcon is true, but size is not defined in props.
Therefor size is not type-safe. People might not know they could change
the size of the icon or what the available sizes are.
Separated Props into 2 different interfaces, one with showIcon true
which has an optional size, the other with optional showIcon false
where size is not allowed.

Closes #334