improvement: withJarb errors in tooltip
If you want to display a form in an EpicTable, showing errors below the
inputs will clutter the frame. Instead, you could display the errors in
a tooltip that'll show on hover.
Added the possibility to specify if the errors should be displayed in a
tooltip or not. Defaults to displaying the errors underneath the input.

Closes #358