fix: ModalPicker should not overflow a fixed-width container
When ModalPickerSingle or ModalPickerMultiple is used within a
fixed-width container like a table cell, a long label should not cause
the button to be (partly) hidden.
Replaced MoreOrLess in ModalPickerMultiple with just a comma separated
list of selected values.
Added an ellipsis when the label is too long.
Added a tooltip to display the full label when ellipsis is used.

Closes #365