improvement: ModalPicker button alignment
You should have the option to align the button used to open the
ModalPicker modal to bring more order in a layout with multiple
ModalPickers (i.e. in forms in EpicTable).
Added an option to ModalPickerOpener to define what position the button
should be displayed. Possible options are 'left', 'right' and 'default',
where 'left' displays the button on the left and options right next to
it, 'right' displays the options on the left and the button on the
right, and 'default' displays the selected values left when there are
any and the button right next to it.

Closes #367