improvement: added possibility to copy/paste to form in EpicTable
To improve user experience, you could add a button "Copy for Excel"
which copies all content in an Excel friendly format.
It would be nice if you could also copy a table in Excel and paste it in
the EpicTable form, so there's no need for an "Import Excel" or "Import
CSV" button.
Added preview of how copy and paste functionality should be coded in
EpicForm stories.

Closes #361