improvement: withJarb support accepting errors as React.ReactNode
This way, the user can render errors in a custom element, for example as
a list item.
Replaced string with ReactNode as type of MetaError to accept custom
nodes as well.
Used the index as key when mapping ReactNode errors, due to which
keyForError function becomes unused, so that function is removed.
Added a preview to Example story first name and last name validation.

Closes #369