fix: duplicate ID in asyncPage stories
The console gives a warning about a duplicate ID being used for stories
in AsyncPage. Also, the duplicate ID caused one of the stories not to be
displayed in the menu.
Changed the ID for one of the stories so there is no warning anymore and
both of the stories are displayed in the menu.