improvement: DateTimeInput datepicker in modal
When used in EpicTable, the DateTimeInput causes extra blank space
underneath the content of the table, so the content will become
scrollable vertically. It should be possible to display the date picker
in a modal. That way, the date picker doesn't conflict visually when
used in a table.
Added an option to display a button next to the inpit that will open the
datepicker in a modal when clicked.

Closes #382