fix: modal closed when valuepicker total elements narrowed by search
When you have a ModalPickerMultiple component with a lot of options and you try to narrow the number of options down by using the search option, the modal closes sometimes.
This seems to happen when you narrow it down too much.

The cause of this problem lies in the containing ValuePicker component.
This component wraps several other components and displays the right one
based on the total options. When the ModalPicker narrows the number of
options, the value of the total options in the ValuePicker was also
changed. The ValuePicker wraps the given optionFetcherCallback in its
own callback function. In this function it updates the total number of
options. And this callback function is passed to its child components.

Moved setting the total elements to the boot function to prevent total
elements from being changed by a query.

Closes #412