feat: added custom popover component.
The `Popover` component used tippy.js like the `Tooltip` component does.

Updated `@tippy.js/react` from `3.1.1` to `4.0.2`. It is under a new
name now: `@tippyjs/react` without the dot.

Also added dependency to `tippy.js` to our library so the scss imports
work. It depends the same version as `tippyjs/react` does.

In main.scss moved the `external` includes above our own custom componets
instead of only above the form components.

Due to the upgrade to 4.0.2 there is a breaking changes for our
library which we are going to ignore. The `distance` property of
the `Tooltip` is now only a number instead of a string in the tippy
library. Because our users only use `distance` as a number we will let
this one slide.

Also added the `offset` to the `Tooltip` and added an example of how
this works.

Removed `boundary="viewport"` from `Tooltip` because it is not needed
anymore according to tippy's docs.

Closes: #421