fix: popover not closed when clicked outside
When the isOpen property is used in the Popover, even though hideOnClick
is true, clicking outside the popover doesn't close the popover, because
the isOpen boolean is static and not manageable by the Popover itself.

Replaced hideOnClick with onClickOutside to take control of what happens
when you click outside the popover.

Closes #424