improvement: RadioGroup, and the modalpickers can now clear the value.
The clear behavior in the `ModalPickerMultiple` and `ModalPickerSingle`
acts and looks the same as the `IconPicker` and `ColorPicker`.

The `RadioGroup` renders the clear button above the radio inputs. But
only when the new `canClear` props is set to `true`.

The `ValuePicker`'s `RadioGroup` will set the `canClear` prop to
`true`. This way all `ValuePicker` variants can clear the selection.

Extracted the shared "clear" button into a `TextButton` component which
is also exposed.

Closes #428