fix: update RadioGroup options when options change
In Select and RadioGroup, when the options property is an array, the
options displayed are not updated when the options array changes.
When the options property is an OptionsFetcher that uses an external
value, there is no possibility to reload the options.

Changed behaviour of options when it's an array.
Added watch property to RadioGroup and Select to watch for external
changes that influence the result of the OptionsFetcher and trigger the
fetcher when the value changes.