improvement: OpenCloseModal now has icons and supports a sticky footer
The sticky footer is enabled by default but can be disabled. When
enabled it will make the footer stay with the user, this way the user
can always press the save and cancel buttons.

The save and cancel buttons also have an icon by default now. The
icons can optionally be changed by the user. The default icons are
`save` and `cancel`.

Also added a story on how to properly use a final-form `Form` inside
of the `OpenCloseModal`. It uses the same form as the `FormExample`
which exports it now. Added it on two locations: with the `FormExample`
and the `OpenCloseModal`, so it can be found more easily.

Closes #452