fix: CheckboxMultipleSelect, and ModalPickerMultiple deselect was broken
The deselect broke when using a custom `isOptionEqual` callback. The
idea of the `isOptionEqual` is that you can say that two objects are
equal based on something other than the default reference equality.
The canonical usecase being checking two `id`'s.

In both the `CheckboxMultipleSelect`, and `ModalPickerMultiple` when
clicking an option a routine is executed when an option is selected,
which should remove it when it is already selected.

In both cases de deselection uses a filter:

```ts
   selected = selected.filter(v: T) => v !== item);
```

Of course this filter uses simple reference equality. Which in cases
when using `isOptionEqual` breaks, because the user will provide us with
different `objects` each time, which of course breaks selection.

The fix is using the `isOptionEqual` instead, simple enough.

However how did this bug exist in the first place. We use objects
everywere in our test and stories? The answer is that the fixtures
are frozen objects. This makes reference equality work... by accident.

Changed all fixtures to be functions instead, this way we always create
new instances, and then these bugs are caught more easily. Of course
this meant changing the code which use the fixtures.

Closes: #447