improvement: required mark in label
When using a Jarb field, the field should know if it's required, so it
should be possible to add a required mark to the label when the field is
required.

Added check that adds required mark to label in Jarb fields when the
field is required.
Added configure function to optionally disable required mark entirely.

Closes #465