fix: EpicResize did not work when columns are dynamically added
When columns are added and removed the resizing of a column will
randomly resize another previously existing column.

The reason for this is because the `EpicResize` uses a throttle:

```ts
const throttledResize = useRef(throttle(onResize, 40));
```

But when the `onResize` changes the ref `throttledResize` is never
updated.

When adding new columns React will reuse the `EpicResize` and this will
of course cause troubles because the `throttledResize` will still
reference the old `onResize`.

By adding a `useEffect` which when the `onResize` changes sets the
`throttledResize` again fixes the issue.

Fixes #471