fix: ImageUpload no longer swallows events because of a high z-index
When used in combination with a `TextEditor` component with a custom
toolbar, the items of the toolbar cannot be clicked when they hover
over the ImageUpload.

The reason for this is the z-index of the ImageUpload being really high.
Therefore getting the click first.

The solution was setting the z-index to zero. This does not seem to
affect anything.

Fixes: #476