fix: ImageUpload replace file extension to match output
ImageUpload accepts any image, but the output is always a PNG file. The
output file doesn't have the .png extension though. This is not a
problem for browsers, as they don't use the extension, but the
extension should match the contents.

Added a function that replaces the file extension with .png.
Added an optional property to keep the original file name.

Closes #480