fix: DateTimeInput cleared input on backspace.
The reason is that the `DateTimeInput` when the input is not a date
sends `null` via `onChange` callback. This causes the `value` to
prop to become `null`. The `null` will then result in an empty input.

To fix this every time the value changes we store a string representation
of that value inside called `lastStringValue`. This value is then given
to the `DateTime` component the `value` prop when the `value` is not
defined. This way the "intermediate" invalid `value` is stored and not
forgotten when the date becomes invalid.

Added some more stories for the `jarb` variant, so we can see that this
behavior is working.

Tiny fix: also added the `id` prop to the `inputProps` of the `DateTime`
component. This way clicking the label focuses the input.

Refactoring: moved all the hooks from the `DateTimeInput` into a hooks
folder.

Fixes #484, and #478