improvement: Pagination component added next and previous buttons.
The `Pagination` will now by default show a next and previous buttons
in the form of an arrow icon. Clicking the buttons will move the the
next or previous page. The previous button is disabled when the user
is already on the first page, the next button is disabled when the
user is already on the last page.

A new property called `showPreviousAndNextButtons` on `Pagination` will
hide the new next and previous button when `showPreviousAndNextButtons`
is `false`.

Changed the `Pagination` components story so it actually updated the
page instead of loggin an action.

Also added documentation on the `pageOf` utility.

Removed nested `Text` prop from the `Pager` this was probably a copy
paste error.

Closes #486