feat: Added ConfirmModal to trigger user confirmation without button
Added `ConfirmModal`, which is based on `OpenCloseModal` and offers
the user a confirmation before performing an action.

A confirmation dialog was previously only accessible through
`ConfirmButton`. While `ConfirmButton` remains the primary choice to
perform operations that need confirmation from the user, in some cases
it could be needed to offer a confirmation modal through other means.

An example is a `ButtonDropdown` which contains a remove action that
needs confirmation. In such case, we want to use the generic confirm
modal but don't want to render a button.