bugfix: TextEditor now longer allows all formatting such as images
Quill separates the toolbar from the allowed formats. For example it is
possible to not show the "bold" toolbar item, but to allow the ctrl-b
shortcut and pasting bold text. The `format` property defines which
formatting is actually allowed.

We did not set the `formats` for the `TextEditor` which then defaults
to allow all formats. This is unexpected, developers do not expect that
for the default toolbar we provide, that the user can for example can
paste images.

The solution is adding a `formats` property to the `TextEditor` which
by default is set allow to all items in the toolbar. This will work for
99% of the use cases in which the developer expects the toolbar define
the formats.

Added a story called 'manual formats` to show that with the `formats`
option you can control the `formats` manually.

Also renamed the old `custom toolbar` story to `custom toolbar option`
because it shows of how to add a custom toolbar option. Added a new story
called `custom toolbar` which shows all toolbar options.

Tiny fix: gave `AsyncContent` tests a mock for the console.error so it
does not show up in the test output.

Closes #490