bugfix: DateTimeInput no longer uses browsers timezone.
When the user picked a date the timezone would be in the timezone that
the browser is currently in. When the user selected `28-10-2019` from
the `DateTimeInput` using the picker the value of the date would become
`2019-10-27T23:00:00.000Z`.

The `2019-10-27T23:00:00.000Z` is parsed by Spring for a `LocalDate`
as `2019-10-27`.

By using setting `react-datetime`'s `utc` property to `true` by default
the date would become: `2019-10-28T00:00:00.000Z` which is correctly
parsed.

By default the `utc` property in the `DateTimeInput` will be `true`
and this is passed to `DateTimeModal` as well.

Updated:

1. `moment` from `2.24.0` to `2.29.1`.

2. `react-datetime` from `2.16.3` to `3.04`. This changes `onFocus` to
   `onOpen` as a breaking change for us. Also the name of the default
    export changed to a strange string, so now in the tests we find
    `Datetime` directly instead of the string variant.

Also exposing the `AutoSave` component in the index export.

Fixes #492